+++ dds-commander/src/UIChannelInfo.cpp
@@ -5,12 +5,19 @@
 
 #include "UIChannelInfo.h"
 #include "Intercom.h"
+#include "UserDefaults.h"
+#include "submit_info.pb.h"
+#include "submit_info_slurm.pb.h"
+#include <boost/filesystem.hpp>
 
 using namespace std;
 using namespace dds;
 using namespace dds::commander_cmd;
+using namespace dds::protocol;
 using namespace dds::protocol_api;
 using namespace dds::misc;
+
+const LPCSTR g_submitInfoFile = "submit.inf";
 
 CSubmitAgentsChannelInfo::CSubmitAgentsChannelInfo()
     : CUIChannelInfo<CSubmitAgentsChannelInfo>()
@@ -57,6 +64,7 @@ bool CSubmitAgentsChannelInfo::processCustomCommandMessage(const SCustomCmdCmd& 
     {
         lock_guard<mutex> lock(m_mutexReceive);
 
+        // Early exit checks
         if (m_channel.expired())
         {
             LOG(fatal) << "Can't process RMS plug-in messages. UI channel expired";
@@ -68,18 +76,12 @@ bool CSubmitAgentsChannelInfo::processCustomCommandMessage(const SCustomCmdCmd& 
             return true;
         }
 
-        // lock UI channel
+        // lock channels
         auto pUI = m_channel.lock();
-        if (!pUI)
-        {
-            LOG(fatal) << "Can't process RMS plug-in messages. UI channel can't be locked";
-            return true;
-        }
-        // lock Plug-in channel
         auto pPlugin = m_channelSubmitPlugin.lock();
-        if (!pPlugin)
+        if (!pUI || !pPlugin)
         {
-            LOG(fatal) << "Can't process RMS plug-in messages. Plug-in channel can't be locked";
+            LOG(fatal) << "Can't process RMS plug-in messages. Channel locks failed";
             return true;
         }
 
@@ -98,8 +100,6 @@ bool CSubmitAgentsChannelInfo::processCustomCommandMessage(const SCustomCmdCmd& 
                     [this](const SSenderInfo& /*_sender*/)
                     {
                         LOG(info) << "Plug-in disconnect subscription called";
-                        // the plug-in is done and went offline, let's close UI
-                        // connection as well.
                         m_channelSubmitPlugin.reset();
                         shutdown();
                     });
@@ -114,6 +114,10 @@ bool CSubmitAgentsChannelInfo::processCustomCommandMessage(const SCustomCmdCmd& 
 
                 intercom_api::SInit init;
                 init.fromPT(pt);
+
+                // Store submission ID from the init message
+                m_sSubmissionID = init.m_id;
+
                 // Sending Submit request to the plug-in
                 SCustomCmdCmd cmd;
                 cmd.m_sCmd = m_strInitialSubmitRequest;
@@ -131,29 +135,76 @@ bool CSubmitAgentsChannelInfo::processCustomCommandMessage(const SCustomCmdCmd& 
                                    ? dds::intercom_api::EMsgSeverity::info
                                    : dds::intercom_api::EMsgSeverity::error));
 
-                // Drop the connection if received an error message from plug-in
+                // FIXME: This is a temporary solution.
+                // When we receive "DDS agents have been submitted" message, we parse submit.inf file
+                if (message.m_msg == "DDS agents have been submitted")
+                {
+                    fs::path pathWorkDirLocalFiles(
+                        smart_path(user_defaults_api::CUserDefaults::instance().getValueForKey("server.work_dir")));
+                    pathWorkDirLocalFiles /= m_sSubmissionID; // Use submission ID instead of channel ID
+                    fs::path pathSubmitInfo(pathWorkDirLocalFiles);
+                    pathSubmitInfo /= g_submitInfoFile;
+
+                    if (fs::exists(pathSubmitInfo))
+                    {
+                        fstream input(pathSubmitInfo.native(), ios::in | ios::binary);
+                        dds::protocol::SubmitInfo protoSubmitInfo;
+                        if (!protoSubmitInfo.ParseFromIstream(&input))
+                        {
+                            throw runtime_error("Failed to parse submission info");
+                        }
+
+                        // Read RMS specific data
+                        dds::protocol::SlurmSubmitInfo protoSlurmSubmitInfo;
+                        if (protoSubmitInfo.rms_plugin_data().UnpackTo(&protoSlurmSubmitInfo))
+                        {
+                            tools_api::SSubmitResponseData submitResponse;
+                            submitResponse.m_requestID = m_requestID;
+
+                            // Copy all job IDs from the protobuf repeated field
+                            for (int i = 0; i < protoSlurmSubmitInfo.slurm_job_id_size(); ++i)
+                            {
+                                submitResponse.m_jobIDs.push_back(protoSlurmSubmitInfo.slurm_job_id(i));
+                            }
+
+                            submitResponse.m_allocNodes = protoSlurmSubmitInfo.alloc_nodes();
+                            submitResponse.m_state = protoSlurmSubmitInfo.state();
+                            submitResponse.m_jobInfoAvailable = protoSlurmSubmitInfo.job_info_available();
+
+                            // Add logging to ensure the response is being sent
+                            LOG(info) << "Sending submit response to UI...";
+                            // Send response via custom command
+                            SCustomCmdCmd cmd;
+                            cmd.m_sCmd = submitResponse.toJSON();
+                            cmd.m_sCondition = "";
+                            pUI->pushMsg<cmdCUSTOM_CMD>(cmd);
+                        }
+                    }
+                }
+
                 if (message.m_msgSeverity == dds::intercom_api::EMsgSeverity::error)
                 {
                     doneWithUI();
                     m_channel.reset();
                 }
             }
+            else if (tag == "submit")
+            {
+                doneWithUI();
+            }
         }
     }
-    // catch explicitly ptree_error. On some systems the exact exception message is not propagated to
-    // exception parent and can be lost
     catch (const boost::property_tree::ptree_error& _e)
     {
         string msg("Failed to process RMS plug-in message: ");
         msg += _e.what();
         LOG(error) << msg;
-
         sendUIMessage(msg, dds::intercom_api::EMsgSeverity::error);
         shutdown();
     }
     catch (const exception& _e)
     {
-        LOG(error) << "CSubmitAgentsChannelInfo::processCustomCommandMessage: ";
+        LOG(error) << "CSubmitAgentsChannelInfo::processCustomCommandMessage: " << _e.what();
         shutdown();
     }
     return true;