+++ dds-misc-lib/src/Res.h
@@ -25,5 +25,5 @@ namespace dds::misc
     const LPCSTR g_cszReportBugsAddr("Report bugs/comments to fairroot@gsi.de");
     const LPCSTR g_cszDDSServerIsNotFound_StartIt("Looks like we can't find any suitable DDS commander server to "
                                                   "connect. Use \"dds-session start\" to start one.");
-};     // namespace dds::misc
+}; // namespace dds::misc
 #endif // _DDS_RES_H_